Code refactor

+++ service/FolderCleanup.py
@@ -2,12 +2,17 @@
 import os
 import shutil
 from dataclasses import dataclass
+from logging import Logger
+from typing import List
+from apscheduler.schedulers.blocking import BlockingScheduler
+
+from common.utils import get_tag, get_formatted_emby, get_formatted_plex, build_target_string
+from common.utils_server import get_connection_info
+
+from service.ServiceBase import ServiceBase
 
 from api.plex import PlexAPI
 from api.emby import EmbyAPI
-from common.types import CronInfo
-from common.utils import get_tag, get_formatted_emby, get_formatted_plex, build_target_string
-from service.ServiceBase import ServiceBase
 
 @dataclass
 class PathInfo:
@@ -17,36 +22,26 @@ class PathInfo:
     emby_library_id: str
 
 class FolderCleanup(ServiceBase):
-    def __init__(self, ansi_code, plex_api, emby_api, config, logger, scheduler):
+    def __init__(self, ansi_code: str, plex_api: PlexAPI, emby_api: EmbyAPI, config, logger: Logger, scheduler: BlockingScheduler):
         super().__init__(ansi_code, self.__module__, config, logger, scheduler)
         
         self.plex_api = plex_api
         self.emby_api = emby_api
-        self.paths = []
-        self.ignore_folder_in_empty_check = []
-        self.ignore_file_in_empty_check = []
+        self.paths: list[PathInfo] = []
+        self.ignore_folder_in_empty_check: list[str] = []
+        self.ignore_file_in_empty_check: list[str] = []
         
         try:
             for path in config['paths_to_check']:
                 plex_library_name = ''
                 if 'plex_library_name' in path:
-                    if self.plex_api.get_valid() == True:
-                        plex_library_name = path['plex_library_name']
-                    else:
-                        self.log_warning('{} library defined but API not valid {} {}'.format(get_formatted_plex(), get_tag('library', path['plex_library_name']), get_tag('plex_valid', self.plex_api.get_valid())))
+                    plex_library_name = path['plex_library_name']
                 
                 emby_library_name = ''
-                emby_library_id = ''
                 if 'emby_library_name' in path:
-                    if self.emby_api.get_valid() == True:
-                        emby_library = self.emby_api.get_library_from_name(path['emby_library_name'])
-                        if emby_library != self.emby_api.get_invalid_item_id():
-                            emby_library_name = path['emby_library_name']
-                            emby_library_id = emby_library['Id']
-                    else:
-                        self.log_warning('{} library defined but API not valid {} {}'.format(get_formatted_emby(), get_tag('library', path['emby_library_name']), get_tag('plex_valid', self.emby_api.get_valid())))
+                    emby_library_name = path['emby_library_name']
                 
-                self.paths.append(PathInfo(path['path'], plex_library_name, emby_library_name, emby_library_id))
+                self.paths.append(PathInfo(path['path'], plex_library_name, emby_library_name, ''))
                 
             for folder in config['ignore_folder_in_empty_check']:
                 self.ignore_folder_in_empty_check.append(folder['ignore_folder'])
@@ -57,7 +52,7 @@ class FolderCleanup(ServiceBase):
         except Exception as e:
             self.log_error('Read config {}'.format(get_tag('error', e)))
 
-    def is_dir_empty(self, dirnames):
+    def is_dir_empty(self, dirnames: List[str]) -> bool:
         dir_empty = True
         for dirname in dirnames:
             if len(self.ignore_folder_in_empty_check) > 0:
@@ -72,7 +67,7 @@ class FolderCleanup(ServiceBase):
                 break
         return dir_empty
     
-    def is_files_empty(self, filenames):
+    def is_files_empty(self, filenames: List[str]) -> bool:
         filenames_empty = True
         for filename in filenames:
             if len(self.ignore_file_in_empty_check) > 0:
@@ -88,25 +83,32 @@ class FolderCleanup(ServiceBase):
         return filenames_empty
     
     def check_delete_empty_folders(self):
-        deleted_paths = []
+        deleted_paths: list[PathInfo] = []
         for path in self.paths:
-            folders_deleted = False
-            
-            keep_running = True
-            while keep_running == True:
-                keep_running = False
-                for dirpath, dirnames, filenames in os.walk(path.path, topdown=False):
-                    if self.is_dir_empty(dirnames) == True and self.is_files_empty(filenames) == True:
-                        self.log_info('Deleting empty {}'.format(get_tag('folder', dirpath)))
-                        shutil.rmtree(dirpath, ignore_errors=True)
-                        keep_running = True
-                        folders_deleted = True
-            
-            if folders_deleted == True:
-                deleted_paths.append(path)
+            connection_info = get_connection_info(self.plex_api, path.plex_library_name, self.emby_api, path.emby_library_name)
+            if (path.plex_library_name == '' or connection_info.plex_valid == True) and (path.emby_library_name == '' or connection_info.emby_valid == True):
+                folders_deleted = False
+
+                keep_running = True
+                while keep_running == True:
+                    keep_running = False
+                    for dirpath, dirnames, filenames in os.walk(path.path, topdown=False):
+                        if self.is_dir_empty(dirnames) == True and self.is_files_empty(filenames) == True:
+                            self.log_info('Deleting empty {}'.format(get_tag('folder', dirpath)))
+                            shutil.rmtree(dirpath, ignore_errors=True)
+                            keep_running = True
+                            folders_deleted = True
+
+                if folders_deleted == True:
+                    deleted_paths.append(PathInfo(path.path, path.plex_library_name, path.emby_library_name, connection_info.emby_library_id))
+            else:
+                if path.plex_library != '' and connection_info.plex_valid == False:
+                    self.log_warning(self.plex_api.get_connection_error_log())
+                if path.emby_library_name != '' and connection_info.emby_valid == False:
+                    self.log_warning(self.emby_api.get_connection_error_log())
         
         for deleted_path in deleted_paths:
-            target_name = ''
+            target_name: str = ''
             if deleted_path.plex_library_name != '':
                 self.plex_api.switch_plex_account_admin()
                 self.plex_api.set_library_scan(deleted_path.plex_library_name)
